requirements.txt
discord.py==2.3.0
aiohttp==3.8.4
aiofiles==23.2.0
asyncio==3.4.3
python-dotenv==1.0.0
colorama==0.4.6
requests==2.31.0
ujson==5.8.0
matplotlib==3.7.1